blackjack_project/requirements.txt
Django==4.2.7
django-cors-headers==4.3.1
mysql-connector-python==8.2.0
python-dotenv==1.0.0
bcrypt==4.1.1
Werkzeug==3.0.1
